mod: create date_time module
Moved all of the conversions related to date and time
to a separate module - date_time.rs
