+++ scylla-rust-wrapper/src/lib.rs
@@ -14,6 +14,7 @@ pub mod cass_error;
 pub mod cass_types;
 pub mod cluster;
 pub mod collection;
+pub mod date_time;
 pub mod exec_profile;
 mod external;
 pub mod future;
